refactor(mcp): extract resources domain
